src/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
arrow==1.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
circuitbreaker==1.4.0
click==8.0.4
cryptography==42.0.7
dataclasses-json==0.6.6
dill==0.3.8
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
jmespath==0.10.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.0
langchain-core==0.2.1
langchain-text-splitters==0.2.0
langsmith==0.1.60
marshmallow==3.21.2
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
numpy==1.26.4
oci==2.127.0
oci-cli==3.42.0
oracledb==2.2.1
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pluggy==1.5.0
prompt-toolkit==3.0.43
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pyOpenSSL==24.1.0
pytest==8.2.1
pytest-warnings==0.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
tenacity==8.3.0
terminaltables==3.1.10
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.9.4
